migration_policia_nacional/requirements.txt
supabase==2.10.0
python-dotenv==1.0.0
tqdm==4.66.1
psycopg2-binary==2.9.9